change input type validation in add record dialog

+++ app/src/main/java/com/abarska/trackensuretest/entities/FuelingAct.kt
@@ -39,7 +39,7 @@ data class FuelingAct(
 
     @NonNull
     @ColumnInfo(name = NUMBER_OF_LITERS)
-    val numberOfLiters: Double,
+    val numberOfLiters: Int,
 
     @NonNull
     @ColumnInfo(name = TOTAL_SPEND)
